requirements/base.txt
CairoSVG==2.5.2
Pillow==9.2.0
PyPDF2==1.28.4
PyYAML==6.0
Whoosh==2.7.4
bleach==4.1.0
celery==5.2.3
django-activity-stream==1.4.0
django-auth-ldap==4.0.0
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-formtools==2.3
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-qsstats-magic==1.1.0
django-solo==2.0.0
django-stronghold==0.4.0
django-widget-tweaks==1.4.12
django-redis==5.4.0
channels==3.0.5
channels-redis==4.1.0
# channels==3.0.5 requires daphne<4
daphne==3.0.2
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
drf-spectacular==0.27.2
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
extract-msg==0.34.3
flanker==0.9.11
flex==6.14.1
furl==2.1.3
fusepy==3.0.1
gevent==21.12.0
geoip2==4.8.0
graphviz==0.17
gunicorn==20.1.0
jsonschema==4.4.0
node-semver==0.8.1
pycountry==22.3.5
pycryptodome==3.10.4
pyotp==2.6.0
python-dateutil==2.8.2
python-magic==0.4.26
python_gnupg==0.4.8
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
django-ratelimit==4.1.0
django-prometheus==2.3.1
locust==2.24.1
reportlab==4.0.9
sendgrid==6.11.0
mailchimp3==3.0.21
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
sentry-sdk==1.5.8
sh==1.14.2
swagger-spec-validator==2.7.4
whitenoise==6.0.0
